requirements.txt
streamlit
moviepy
langgraph
langchain 
langchain-core 
langchain-google-genai 
google-generativeai
apscheduler
mysql-connector-python
python-decouple